[REF] spreadsheet: split `Spreadsheet` component
This commit extract the logic that displays a spreadsheet into a new
`SpreadsheetEditor` component, only keeping the main spreadsheet/dashboard
and `env` logic in the `Spreadsheet` component.

Task: 4497373
