+++ src/components/spreadsheet/spreadsheet.xml
@@ -1,39 +1,12 @@
 <templates>
   <t t-name="o-spreadsheet-Spreadsheet">
-    <div
-      class="o-spreadsheet h-100 w-100"
-      t-att-class="{'o-spreadsheet-mobile': env.isSmall}"
-      t-ref="spreadsheet"
-      t-att-style="getStyle()">
+    <div class="o-spreadsheet h-100 w-100" t-att-class="{'o-spreadsheet-mobile': env.isSmall}">
       <t t-if="env.isDashboard()">
-        <SpreadsheetDashboard getGridSize.bind="getGridSize"/>
+        <SpreadsheetDashboard/>
         <FullScreenChart/>
       </t>
       <t t-else="">
-        <div class="o-spreadsheet-topbar-wrapper o-two-columns">
-          <TopBar onClick="() => this.focusGrid()" dropdownMaxHeight="gridHeight"/>
-        </div>
-        <div
-          class="o-grid-container"
-          t-att-class="{'o-two-columns': !sidePanel.isMainPanelOpen}"
-          t-att-style="gridContainerStyle"
-          t-on-click="this.focusGrid">
-          <div class="o-top-left"/>
-          <div class="o-column-groups">
-            <HeaderGroupContainer layers="colLayers" dimension="'COL'"/>
-          </div>
-          <div class="o-row-groups">
-            <HeaderGroupContainer layers="rowLayers" dimension="'ROW'"/>
-          </div>
-          <div class="o-group-grid overflow-hidden">
-            <Grid exposeFocus="(focus) => this._focusGrid = focus" getGridSize.bind="getGridSize"/>
-          </div>
-        </div>
-        <SidePanels/>
-        <div class="o-spreadsheet-bottombar-wrapper o-two-columns overflow-hidden">
-          <SmallBottomBar t-if="env.isSmall" onClick="() => this.focusGrid()"/>
-          <BottomBar t-else="" onClick="() => this.focusGrid()"/>
-        </div>
+        <SpreadsheetEditor/>
       </t>
     </div>
   </t>
