+++ tests/grid/grid_component.test.ts
@@ -1642,7 +1642,7 @@ describe("Events on Grid update viewport correctly", () => {
     });
     // mock a resizing of the grid DOM element. can occur if resizing the browser or opening the sidePanel
     extendMockGetBoundingClientRect({
-      "o-spreadsheet": () => ({ x: 0, y: 0, width: 800, height: 650 }),
+      "o-spreadsheet-editor": () => ({ x: 0, y: 0, width: 800, height: 650 }),
       "o-grid": () => ({ x: 0, y: 0, width: 800, height: 650 }),
     });
     // force a triggering of all resizeObservers to ensure the grid is resized
